pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tutorial-vm"
version = "0.0.1a0"
authors = [
  { name="Liezl Maree", email="lmaree@salk.edu" },
  { name="Andrew Park", email="apark@salk.edu" },
  { name= "Elizabeth Berrigan", email="eberrigan@salk.edu" },
]
description = "Scripts for VM instances of tutorial platform."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
    "google-cloud-logging",
    "google-cloud-pubsub",
    "google-cloud-spanner==3.42.0",
    "psutil",
    "sleap[pypi]==1.3.4"
]
keywords = ["tutorial", "vm", "sleap", "google-cloud"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha"
]

[project.urls]
Homepage = "https://github.com/talmolab/cosyne_vm"
Issues = "https://github.com/talmolab/cosyne_vm/issues"

[project.optional-dependencies]
dev = [
    "toml",
    "twine",
    "build"
]

[project.scripts]